# server/swagger.yaml
openapi: 3.0.3
info:
  title: TochkaLi API
  description: |
    API документация для приложения TochkaLi
    
    ## Структура API
    
    API разделен на две части:
    - **Client** - эндпоинты для клиентского приложения (Telegram Web App)
    - **Admin** - эндпоинты для админ панели
    
    ⚠️ **Примечание:** authMiddleware и requireAdmin НЕ используются, все эндпоинты публичные.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: /api
    description: API сервер
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
...
    Broadcast:
      type: object
      properties:
        _id:
          type: string
        title:
          type: string
          example: "Приветственная рассылка"
          description: Название рассылки (уникальное)
        imgUrl:
          type: string
          nullable: true
          description: URL изображения
        content:
          type: string
          description: Содержание рассылки (HTML)
        buttonText:
          type: string
          nullable: true
          description: Текст кнопки
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ==================== Horoscope схемы ====================
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен авторизации
paths:
...
  /user/logout:
    post:
      tags: [Client - Auth]
      summary: Выйти из аккаунта
      description: Завершает активную сессию и инвалидирует refresh token
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Сессия завершена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # ============================================================
  # КЛИЕНТСКАЯ ЧАСТЬ - PROFILE
  # ============================================================
...
  /dynamic-content:
    get:
      tags: [Admin - DynamicContent]
      summary: Весь динамический контент
      responses:
        '200':
          description: Список

    post:
      tags: [Admin - DynamicContent]
      summary: Создать контент
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - content
              properties:
                name:
                  type: string
                content:
                  type: string
      responses:
        '201':
          description: Создан
...
  /broadcast/send:
    post:
      tags: [Admin - Broadcast]
      summary: Отправить рассылку
      description: |
        Отправляет рассылку пользователям. Можно использовать сохраненную рассылку через broadcastId или broadcastTitle,
        либо указать параметры напрямую (message, imageUrl, buttonText и т.д.)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                  description: HTML текст сообщения (необязательно, если указан broadcastId или broadcastTitle)
                broadcastId:
                  type: string
                  description: ID сохраненной рассылки для использования
                broadcastTitle:
                  type: string
                  description: Название сохраненной рассылки для использования
                status:
                  type: string
                  enum: [all, guest, registered, active, client]
                  description: Фильтр по статусу пользователей
                search:
                  type: string
                  description: Поиск по имени, username, телефону
                userIds:
                  type: array
                  items:
                    type: string
                  description: Массив ID пользователей для выборочной отправки
                imageUrl:
                  type: string
                  description: URL изображения для рассылки
                parseMode:
                  type: string
                  enum: [HTML, Markdown]
                  default: HTML
                  description: Режим форматирования текста
                buttonText:
                  type: string
                  description: Текст кнопки
                buttonUrl:
                  type: string
                  description: URL кнопки (будет автоматически сформирован с параметрами пользователя)
      responses:
        '200':
          description: Результат рассылки
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  sent:
                    type: number
                  failed:
                    type: number
                  total:
                    type: number
                  failedUsers:
                    type: array
                    items:
                      type: object
        '400':
          description: Ошибка валидации
        '404':
          description: Сохраненная рассылка не найдена
  /broadcast/test:
    post:
      tags: [Admin - Broadcast]
      summary: Отправить тестовое сообщение
      description: Отправляет тестовое сообщение текущему пользователю (админу)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
                  description: Текст тестового сообщения
      responses:
        '200':
          description: Сообщение отправлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Нет telegramId или сообщение не указано
  /broadcast:
    post:
      tags: [Admin - Broadcast]
      summary: Создать сохраненную рассылку
      description: Сохраняет рассылку для последующего переиспользования
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - content
              properties:
                title:
                  type: string
                  description: Название рассылки (уникальное)
                  example: "Приветственная рассылка"
                imgUrl:
                  type: string
                  description: URL изображения
                content:
                  type: string
                  description: Содержание рассылки (HTML)
                buttonText:
                  type: string
                  description: Текст кнопки
      responses:
        '201':
          description: Рассылка создана
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Broadcast'
        '400':
          description: Ошибка валидации или рассылка с таким названием уже существует

    get:
      tags: [Admin - Broadcast]
      summary: Получить все сохраненные рассылки
      description: Возвращает список всех сохраненных рассылок
      responses:
        '200':
          description: Список рассылок
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Broadcast'
                  count:
                    type: number
...
  /upload/image:
    post:
      tags: [Admin - Upload]
      summary: Загрузить изображение
      description: Max 5MB. JPEG, PNG, GIF, WEBP
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Загружено
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  imageUrl:
                    type: string
                  filename:
                    type: string
        '400':
          description: Ошибка загрузки
  /upload/delete:
    post:
      tags: [Admin - Upload]
      summary: Удалить изображение
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - filename
              properties:
                filename:
                  type: string
      responses:
        '200':
          description: Удалено
        '404':
          description: Не найден

  # ============================================================
  # СИСТЕМНЫЕ - ROBOKASSA
  # ============================================================
...